+++ out/NodeStatus.js
@@ -11,8 +11,9 @@ class NodeStatus {
     }
     update(nodeRoot) {
         const exists = this.pathExists(path.join(`${nodeRoot}`, 'package.json'));
-        this.statusBarItem.text = exists ? '$(check) Node View' : '$(cross) Node View';
+        this.statusBarItem.text = exists ? '$(devspace-check) Node View' : '$(devspace-cross) Node View';
         this.statusBarItem.tooltip = exists ? `Node Root: ${nodeRoot}` : 'No Node Root';
+        this.statusBarItem.command = 'devspace.showNodeView';
     }
     getStatusBarItem() {
         return this.statusBarItem;