+++ src/NodeStatus.ts
@@ -13,8 +13,9 @@ export class NodeStatus {
 
     update(nodeRoot: string | undefined) {
         const exists = this.pathExists(path.join(`${nodeRoot}`, 'package.json'));
-        this.statusBarItem.text = exists ? '$(check) Node View' : '$(cross) Node View';
+        this.statusBarItem.text = exists ? '$(devspace-check) Node View' : '$(devspace-cross) Node View';
         this.statusBarItem.tooltip = exists ? `Node Root: ${nodeRoot}` : 'No Node Root';
+        this.statusBarItem.command = 'devspace.showNodeView';
     }
 
     getStatusBarItem() {